Move it sp_inherents instead
